requirements.txt
Flask==2.0.3
nltk==3.7
numpy==1.22.3
PyMuPDF==1.19.6
python-docx == 0.8.11
matplotlib
wordcloud